Tilgang for å opprette porteføljeoversikt-visninger er nå basert på tilgang til listen
